refactor: rewrite e2e tests in vanilla JS
To save ourselves a build-step and make it easier to debug.
