+++ vscode-extension/test/e2e/suite/completion/completion.test.js
@@ -1,5 +1,5 @@
-import { getDocUri, showFile, position, sleepCI } from "../util";
-import { testCompletion } from "./helper";
+const { getDocUri, showFile, position, sleepCI } = require("../util");
+const { testCompletion } = require("./helper");
 
 describe("SCSS Completion Test", function () {
 	const docUri = getDocUri("completion/main.scss");
@@ -312,8 +312,8 @@ describe("SassDoc Completion Test", () => {
 	it("Offers completions for SassDoc block on mixin without parameters or @content", async () => {
 		const expectedCompletions = [
 			{
-				label: "SassDoc block",
-				insertText: '" ${0}\\n/// @output ${2}"',
+				label: "SassDoc Block",
+				insertText: '" ${0}\\n/// @output ${1}"',
 			},
 		];
 
@@ -323,7 +323,7 @@ describe("SassDoc Completion Test", () => {
 	it("Offers completions for SassDoc block on mixin with @content", async () => {
 		const expectedCompletions = [
 			{
-				label: "SassDoc block",
+				label: "SassDoc Block",
 				insertText: '" ${0}\\n/// @content ${1}\\n/// @output ${2}"',
 			},
 		];
@@ -334,9 +334,9 @@ describe("SassDoc Completion Test", () => {
 	it("Offers completions for SassDoc block on mixin with parameters", async () => {
 		const expectedCompletions = [
 			{
-				label: "SassDoc block",
+				label: "SassDoc Block",
 				insertText:
-					'" ${0}\\n/// @param {${1:Number}} \\\\$a [1px] ${2:-}\\n/// @param {${3:Number}} \\\\$b [2px] ${4:-}\\n/// @output ${6:-}"',
+					'" ${0}\\n/// @param {${1:Number}} \\\\$a [1px] ${2:-}\\n/// @param {${3:Number}} \\\\$b [2px] ${4:-}\\n/// @output ${5}"',
 			},
 		];
 
@@ -346,9 +346,9 @@ describe("SassDoc Completion Test", () => {
 	it("Offers completions for SassDoc block on mixin with parameters and @content", async () => {
 		const expectedCompletions = [
 			{
-				label: "SassDoc block",
+				label: "SassDoc Block",
 				insertText:
-					'" ${0}\\n/// @param {${1:type}} \\\\$a ${2:-}\\n/// @param {${3:type}} \\\\$b ${4:-}\\n/// @output ${6:-}"',
+					'" ${0}\\n/// @param {${1:type}} \\\\$a ${2:-}\\n/// @param {${3:type}} \\\\$b ${4:-}\\n/// @content ${5}\\n/// @output ${6}"',
 			},
 		];
 
@@ -358,9 +358,8 @@ describe("SassDoc Completion Test", () => {
 	it("Offers completions for SassDoc block on parameterless function", async () => {
 		const expectedCompletions = [
 			{
-				label: "SassDoc block",
-				insertText:
-					'" ${0}\\n/// @param {${1:type}} \\\\ ${2:-}\\n/// @return {${3:type}} ${4:-}"',
+				label: "SassDoc Block",
+				insertText: '" ${0}\\n/// @return {${1:type}} ${2:-}"',
 			},
 		];
 
@@ -370,7 +369,7 @@ describe("SassDoc Completion Test", () => {
 	it("Offers completions for SassDoc block on parameterfull function", async () => {
 		const expectedCompletions = [
 			{
-				label: "SassDoc block",
+				label: "SassDoc Block",
 				insertText:
 					'" ${0}\\n/// @param {${1:Number}} \\\\$a [1px] ${2:-}\\n/// @param {${3:Number}} \\\\$b [2px] ${4:-}\\n/// @return {${5:type}} ${6:-}"',
 			},
