+++ vscode-extension/test/e2e/suite/signature/helper.js
@@ -1,18 +1,22 @@
-import * as assert from "assert";
-import * as vscode from "vscode";
-import { showFile } from "../util";
+const assert = require("assert");
+const vscode = require("vscode");
+const { showFile } = require("../util");
 
-export async function testSignature(
-	docUri: vscode.Uri,
-	position: vscode.Position,
-	signature: vscode.SignatureHelp,
-) {
+/**
+ * @param {import('vscode').Uri} docUri
+ * @param {import('vscode').Position} position
+ * @param {import('vscode').SignatureHelp} signature
+ * @returns {Promise<void>}
+ */
+async function testSignature(docUri, position, signature) {
 	await showFile(docUri);
 
-	const result = await vscode.commands.executeCommand<vscode.SignatureHelp>(
-		"vscode.executeSignatureHelpProvider",
-		docUri,
-		position,
+	const result = /** @type {import('vscode').SignatureHelp} */ (
+		await vscode.commands.executeCommand(
+			"vscode.executeSignatureHelpProvider",
+			docUri,
+			position,
+		)
 	);
 
 	if (result === undefined) {
@@ -52,3 +56,7 @@ export async function testSignature(
 		);
 	});
 }
+
+module.exports = {
+	testSignature,
+};
