+++ vscode-extension/test/fixtures/references/one.scss
