--- vscode-extension/test/web/suite/example.test.ts
@@ -1,11 +1,0 @@
-import * as assert from "assert";
-import * as vscode from "vscode";
-
-describe("Web Extension Test Suite", () => {
-	vscode.window.showInformationMessage("Start all tests.");
-
-	it("Sample test", () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-	});
-});
