+++ vscode-extension/test/web/suite/hover.test.js
@@ -1,18 +1,22 @@
-import * as assert from "assert";
-import * as vscode from "vscode";
-import { getDocUri, showFile, position, sleep } from "./util";
+const assert = require("assert");
+const vscode = require("vscode");
+const { getDocUri, showFile, position, sleep } = require("./util");
 
-async function testHover(
-	docUri: vscode.Uri,
-	position: vscode.Position,
-	expectedHover: vscode.Hover,
-) {
+/**
+ * @param {import('vscode').Uri} docUri
+ * @param {import('vscode').Position} position
+ * @param {import('vscode').Hover} expectedHover
+ * @returns {Promise<void>}
+ */
+async function testHover(docUri, position, expectedHover) {
 	await showFile(docUri);
 
-	const result: any[] = await vscode.commands.executeCommand(
-		"vscode.executeHoverProvider",
-		docUri,
-		position,
+	const result = /** @type {import('vscode').Hover[]} */ (
+		await vscode.commands.executeCommand(
+			"vscode.executeHoverProvider",
+			docUri,
+			position,
+		)
 	);
 
 	if (!result[0]) {
@@ -21,7 +25,9 @@ async function testHover(
 
 	const contents = result
 		.map((item) => {
-			return item.contents.map((content: any) => content.value);
+			return item.contents.map((content) => {
+				return /** @type {import('vscode').MarkdownString} */ (content).value;
+			});
 		})
 		.join("\n");
 
@@ -103,7 +109,7 @@ describe("Hover", () => {
 		// Prefixed symbols are shown with their original names
 		const expectedContents = {
 			contents: [
-				"```scss\n@mixin mix-mixin()\n```\n____\nMixin declared in _mixins.scss",
+				"```scss\n@mixin mix-mix-mixin()\n```\n____\nMixin declared as mix-mixin in _mixins.scss",
 			],
 		};
